Coerce an Object of Class simculta to a Data Frame

Coerce an Object of Class simculta to a Data Frame

Usage

# S3 method for class 'simculta'
as.data.frame(x, ...)

Arguments

x

Object of class simculta.

...

Additional arguments.

Author

Ivan Jacob Agaloos Pesigan

Examples

x <- GenCULTA2Profiles(
  n = 10,
  m = 6,
  common_trait_loading = matrix(
    data = c(1, 1.25, 1.50, 1.75),
    ncol = 1
  ),
  common_state_loading = matrix(
    data = c(1, 1.5, 1.75, 2.00),
    ncol = 1
  ),
  mu_t = NULL,
  psi_t = NULL,
  mu_p = NULL,
  psi_p = NULL,
  theta = diag(4),
  mu_profile = cbind(
    c(-3, -3, -3, -3),
    c(3, 3, 3, 3)
  ),
  mu_x = 0,
  sigma_x = 1,
  nu_0 = -3.563,
  kappa_0 = 0.122,
  alpha_0 = -3.586,
  beta_00 = 2.250,
  gamma_00 = 0.063,
  gamma_10 = 0.094,
  phi_0 = 0.311,
  phi_1 = 0,
  psi_s0 = 0.151,
  psi_s = 0.290
)
as.data.frame(x)
#>    id   covariate      y1t0     y2t0      y3t0       y4t0       y1t1     y2t1
#> 1   1 -0.05788734 5.5586108 4.015409 4.8071006  4.4820997  3.3488802 0.764708
#> 2   2  0.48236947 2.0215636 3.070005 2.1049165 -2.7850668  0.9772593 2.510284
#> 3   3  0.99294364 1.3037647 3.321640 1.4157185  2.0901442 -0.2278127 4.816720
#> 4   4 -1.24639550 0.6134074 2.033937 0.4115373 -1.5233829  1.5612930 3.375865
#> 5   5 -0.03348752 1.8860911 3.783625 4.7466582  3.9078727  3.8195600 2.155750
#> 6   6 -0.07096218 2.7039843 1.850319 1.0970003 -0.6784618  1.9165575 3.991854
#> 7   7 -0.75892065 3.6785650 3.973106 4.9584789  4.8689740  1.4758460 4.059789
#> 8   8 -1.03435936 3.4897155 4.715623 0.2789321  4.3518521  3.9216108 2.833675
#> 9   9 -0.63073195 4.0687524 2.778493 2.6382359  3.6560427  5.7206406 1.246627
#> 10 10  0.58680772 0.8762377 2.592222 3.5173125  5.2557097  3.2282111 1.409625
#>         y3t1          y4t1        y1t2      y2t2       y3t2       y4t2
#> 1  3.2176211  3.1228094771  3.68708162 3.6624330  5.3469261  3.8077072
#> 2  2.0397073 -0.7634759375  1.94784303 1.1031107  2.3692209 -4.2720781
#> 3  1.7584798  2.3211517173 -0.11585323 2.3422468  1.4097085  3.5489662
#> 4  3.6867485 -0.8057827984  0.05430914 1.2606708  0.6998451 -0.5398667
#> 5  2.7454002  4.6403705138  5.52849153 5.5563751  6.0189471  6.3926502
#> 6  3.1266704  1.8386089959  1.36995274 0.6800256  2.2480258 -1.7645419
#> 7  3.1036400  2.7413830032  4.50554804 5.7032839  4.5790373  5.9187316
#> 8  0.4824264  0.0002365731  4.59983665 4.6573894 -0.5586842  3.0691371
#> 9  3.6659849  1.9484396349  4.37019908 1.6968734  2.8650687 -0.1621605
#> 10 2.9555879  3.5332211353  2.97357757 2.0905770  1.7684167  3.5536691
#>         y1t3      y2t3      y3t3       y4t3      y1t4      y2t4       y3t4
#> 1  3.6305037 2.8777504 3.1181619  3.1103786  4.482147 3.3478626  3.7672389
#> 2  1.4842027 2.5274600 5.3132418 -1.2047034  2.311866 0.8845128  3.0944010
#> 3  0.0794906 4.2323747 0.4909875  2.9996484 -0.427624 2.3119721 -2.5677444
#> 4  1.6816458 2.4159691 3.6239146  1.8704750 -1.340042 1.6044595  2.6365820
#> 5  3.3166050 1.8331067 4.2341816  2.0370051  3.409822 3.9350173  3.7235348
#> 6  1.3399947 1.8678224 3.4911126 -1.5826918  1.413342 1.6891299  0.5433248
#> 7  3.0641576 3.4059414 3.5210034  4.9922222  4.289931 5.6630787  4.7003826
#> 8  5.5093757 3.4140741 3.0066723  2.0787009  4.484684 4.1776158  1.8014695
#> 9  2.8859008 0.9954588 3.3837934  0.7397054  4.057643 2.2076692  2.1132591
#> 10 5.5191044 4.0339255 2.3105899  3.6790048  2.501198 5.1153146  3.1940027
#>          y4t4       y1t5     y2t5       y3t5       y4t5
#> 1   5.9622371  4.3622596 2.205020  1.7748956 2.07404394
#> 2  -2.1983096  2.2143933 4.433160  5.8345936 1.51033647
#> 3  -0.6651571  0.3083751 3.753993  0.8472692 1.52210683
#> 4  -2.3444524 -0.2828603 1.857651  1.3447812 0.02520816
#> 5   3.7509303  4.0586701 1.673024  2.4320729 3.53000955
#> 6  -2.4594718  1.1081651 2.514302  3.2423761 0.09296968
#> 7   5.2049353  2.3931431 4.824177  4.7840672 5.54305879
#> 8   2.3653249  5.6682881 4.497550 -0.1351435 1.82779548
#> 9  -1.4275159  4.3498987 3.864494  2.5115426 0.64152713
#> 10  5.4159502  0.6198721 2.734512  2.9141825 3.13509945