Coerce an Object of Class simculta to a Data Frame

Coerce an Object of Class simculta to a Data Frame

Usage

# S3 method for class 'simculta'
as.data.frame(x, ...)

Arguments

x

Object of class simculta.

...

Additional arguments.

Author

Ivan Jacob Agaloos Pesigan

Examples

x <- GenCULTA2Profiles(
  n = 10,
  m = 6,
  common_trait_loading = matrix(
    data = c(1, 1.25, 1.50, 1.75),
    ncol = 1
  ),
  common_state_loading = matrix(
    data = c(1, 1.5, 1.75, 2.00),
    ncol = 1
  ),
  mu_t = NULL,
  psi_t = NULL,
  mu_p = NULL,
  psi_p = NULL,
  theta = diag(4),
  mu_profile = cbind(
    c(-3, -3, -3, -3),
    c(3, 3, 3, 3)
  ),
  mu_x = 0,
  sigma_x = 1,
  nu_0 = -3.563,
  kappa_0 = 0.122,
  alpha_0 = -3.586,
  beta_00 = 2.250,
  gamma_00 = 0.063,
  gamma_10 = 0.094,
  beta_0 = 0.311,
  beta_1 = 0,
  psi_s0 = 0.151,
  psi_s = 0.290
)
as.data.frame(x)
#>    id   covariate       y1t0     y2t0     y3t0       y4t0        y1t1     y2t1
#> 1   1  0.73507214  4.3320838 3.711972 5.463762  3.5787691  5.76427170 5.911547
#> 2   2 -0.14647263  0.4570416 1.971575 1.768068 -0.9125972  2.55672265 5.383108
#> 3   3 -0.05788734  5.0255161 5.370127 3.274071  3.2957101  2.92403909 2.281807
#> 4   4  0.48236947  5.5750629 4.676081 4.257788  3.6646698  4.03601234 3.374241
#> 5   5  0.99294364  5.0312760 2.433218 4.232033  4.2283717  2.50804583 2.440819
#> 6   6 -1.24639550  1.1087698 1.531356 1.527814 -2.3167256  1.46876683 1.991451
#> 7   7 -0.03348752 -1.8047358 3.960042 0.545138  0.2728378 -0.02138524 2.106989
#> 8   8 -0.07096218  2.1770779 2.273373 1.336797  0.5317250  0.79776730 3.527082
#> 9   9 -0.75892065  4.4060303 2.452470 5.841048  6.3360889  3.32601122 4.223203
#> 10 10 -1.03435936  0.3951316 3.056125 1.376092  1.1216052  0.61226940 0.852041
#>         y3t1       y4t1      y1t2      y2t2      y3t2       y4t2       y1t3
#> 1   4.657119  5.4445856 3.7339051 3.7633436 4.3308693  3.7727702 2.05135638
#> 2   2.039962  0.2113470 1.2364497 3.4463390 0.6516189 -3.3521754 0.78742123
#> 3   1.874036  2.1529273 3.5793442 5.6551874 4.5582724  3.4720324 3.03204460
#> 4   3.326773  4.6967682 6.2585146 3.8449451 5.8471442  3.6920031 4.72230888
#> 5   2.839402  2.4760737 3.0739171 0.6472130 3.2849760  4.6117118 3.96701212
#> 6   3.774220 -2.7749025 0.5350199 0.7361122 1.7904812 -1.3625126 3.01662668
#> 7  -1.718827  0.7050989 1.2940636 4.9173903 0.8661251  1.6704133 0.08191676
#> 8   2.330671  1.8650283 1.2284490 1.6811834 3.1622772  0.2164504 0.59749562
#> 9   5.950934  6.4538978 4.4380614 2.9902203 4.0704404  5.7959427 3.95475553
#> 10  1.203760 -3.6595254 1.2547925 2.6222009 0.1001698 -0.6620303 1.39835128
#>         y2t3     y3t3       y4t3        y1t4     y2t4     y3t4       y4t4
#> 1  1.5290112 2.753057  3.2897523  3.77726339 4.010253 5.513423  3.7907249
#> 2  6.1397522 2.731125 -2.6582915  0.03795059 1.829367 2.202067 -3.6263766
#> 3  4.1344223 1.470745  1.7932604  3.89248102 4.617724 2.135210  4.6627045
#> 4  3.6604463 6.914176  4.6142469  5.58948723 2.076772 4.764487  3.6996715
#> 5  3.5839676 3.587319  5.4578964  4.63878548 3.431897 2.591641  4.1508667
#> 6  3.1728158 6.209523  2.7563695 -0.37763917 1.802440 4.570180 -2.2037133
#> 7  2.6937314 0.830904 -0.6857525  0.90163773 5.885397 1.591638  2.4811797
#> 8  1.9674872 3.353622 -0.8036903  1.37207572 2.840643 1.632991 -0.2780055
#> 9  2.1300046 4.689054  6.7856024  4.39759752 3.212745 4.498303  5.4324523
#> 10 0.2299153 2.325061 -3.1844270  0.57817027 1.300224 1.477752 -2.4887809
#>         y1t5     y2t5       y3t5       y4t5
#> 1  1.9147250 4.636038  7.5794199  4.6587818
#> 2  2.5457471 2.742045  3.3191851  0.1169530
#> 3  3.8606142 3.606912  0.2969028  0.9505530
#> 4  5.2600907 5.277533  7.0988122  6.9444695
#> 5  3.1713268 1.568731  2.1506041  1.9312152
#> 6  0.6660305 2.035363  3.2547325  0.1389226
#> 7  1.2273440 3.138692 -0.2653210  1.6139763
#> 8  0.9797143 3.535039  4.1794702  2.1000678
#> 9  2.6925009 2.661380  4.9174481  6.1539588
#> 10 2.4717744 2.685157  0.7836388 -1.6063111