Coerce an Object of Class simculta to a Matrix

Coerce an Object of Class simculta to a Matrix

Usage

# S3 method for class 'simculta'
as.matrix(x, ...)

Arguments

x

Object of class simculta.

...

Additional arguments.

Author

Ivan Jacob Agaloos Pesigan

Examples

x <- GenCULTA2Profiles(
  n = 10,
  m = 6,
  common_trait_loading = matrix(
    data = c(1, 1.25, 1.50, 1.75),
    ncol = 1
  ),
  common_state_loading = matrix(
    data = c(1, 1.5, 1.75, 2.00),
    ncol = 1
  ),
  mu_t = NULL,
  psi_t = NULL,
  mu_p = NULL,
  psi_p = NULL,
  theta = diag(4),
  mu_profile = cbind(
    c(-3, -3, -3, -3),
    c(3, 3, 3, 3)
  ),
  mu_x = 0,
  sigma_x = 1,
  nu_0 = -3.563,
  kappa_0 = 0.122,
  alpha_0 = -3.586,
  beta_00 = 2.250,
  gamma_00 = 0.063,
  gamma_10 = 0.094,
  phi_0 = 0.311,
  phi_1 = 0,
  psi_s0 = 0.151,
  psi_s = 0.290
)
as.matrix(x)
#>    id   covariate      y1t0      y2t0       y3t0        y4t0       y1t1
#> 1   1  0.96298228  5.169668  4.274674  6.2343811  7.10176919  4.5795000
#> 2   2  0.05357102 -3.179332 -3.078806 -1.9896289 -1.09144570 -4.0662783
#> 3   3 -0.43489841  1.741236  3.664483  0.9649907 -0.06289595 -0.0227294
#> 4   4 -1.73729726  4.843360  3.480369  1.4274894  3.36510490  5.6259881
#> 5   5 -1.26369560  3.782242  1.351715  2.9427369  1.79492136  4.1703608
#> 6   6  0.40630851  2.153713  2.448050  6.4307622  0.82026921  0.9993397
#> 7   7 -1.45965397  3.187589  4.987633  6.4153168  6.54650864  4.8089617
#> 8   8  1.04845737  3.108124  3.289756  6.9083992  5.16560453  4.9094513
#> 9   9 -1.34643054  2.499432  3.046025  0.4712187 -2.38458438  2.7194779
#> 10 10 -0.19357056  0.299359  4.619669  3.4303420  5.35578262  0.6633495
#>          y2t1      y3t1        y4t1      y1t2     y2t2       y3t2       y4t2
#> 1   6.7192306  6.045733  5.15785621 5.3308925 4.480549  6.2758984  5.3035594
#> 2  -3.2791562 -2.302476 -1.52292783 2.4955208 3.238228  5.9966701  5.7236904
#> 3   2.0769196  0.374529 -0.01985354 2.1532250 1.685218  0.2291469  1.1929425
#> 4   4.6009064  4.667957  4.78767319 4.2175993 3.995034  3.0083666  2.2328544
#> 5   1.4074369  1.257834  0.37680413 4.2422077 2.968707  3.5152542  4.5464480
#> 6   4.4365377  4.991256  1.82479659 0.8520359 1.668367  3.8500824 -0.1251810
#> 7   3.8389022  6.213358  5.41862987 2.7582569 3.718020  3.3514878  2.2179556
#> 8   0.5089100  2.108597  4.03151989 5.7788110 3.245956  4.8326252  6.6788243
#> 9   2.5678902 -1.697736  0.77904779 2.5939581 0.540211 -1.2891760 -0.8636479
#> 10  0.6898624  2.618613  0.89160197 2.0970055 3.870696  6.5278311  6.8438020
#>          y1t3       y2t3       y3t3       y4t3      y1t4       y2t4       y3t4
#> 1   4.9647455  4.9061314  5.1533464  3.9193757  5.209965  5.0213801  4.7511168
#> 2   2.7038780  1.1295392  4.6699030  3.7237263  3.863576  2.3339723  2.6999287
#> 3   1.5673056  3.4527936 -1.2532654 -0.1382367 -0.551350  1.0349972 -0.8588542
#> 4  -0.5632622 -0.7069522 -2.7720036 -2.8605303  4.520011  2.5073466  5.2868830
#> 5   4.9475164  1.4066976  1.9458753  1.9402478  3.175947  3.2296663  2.6507675
#> 6   1.3846041  2.0482936  4.1800220  0.4175228  2.406961  1.9072895  3.5058954
#> 7   2.9162976  2.7762897  3.3360606  1.6816822 -1.981660 -0.9395829 -3.3575108
#> 8   3.2281550  0.2785083  5.2362433  4.2728986  1.520190 -0.1393181  3.1058483
#> 9   4.6803712  3.9607754  0.9578368 -1.6505130  3.436325  2.5882202 -1.3595748
#> 10  3.0067244  2.0715573  2.8240631  4.8080380  1.587663  5.7102471  4.9006622
#>          y4t4        y1t5       y2t5      y3t5       y4t5
#> 1   4.2689087  5.76586175  3.5013596  6.056141  5.1135383
#> 2   4.4653561  3.42174181  6.1022666  7.359658  4.4224276
#> 3  -0.9814545 -0.92780702  0.4312681 -2.461801  0.6940986
#> 4   4.6154452  5.77366092  2.6343353  1.759367  1.3968721
#> 5   1.8074246  7.45950081  3.8540892  3.922531  3.6224112
#> 6  -1.4059765  0.02402742  1.8011077  1.785963 -0.2510378
#> 7  -3.1744468 -1.36209878 -0.4938788 -2.126168 -3.1035824
#> 8   4.4357440  4.93773433  0.6300293  1.902562  4.1126967
#> 9  -2.1919886  3.40122956  3.6019252  2.174662  0.4517419
#> 10  4.7603321  0.90799663  2.4760406  2.523084  3.0801351